requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
PyPDF2
sentence-transformers
python-dotenv
streamlit
requests
pdfplumber
langchain-groq
-e .